chore: sonar code smell reduction
