+++ packages/annotation-converter/src/index.ts
@@ -1,1 +1,3 @@
 export * from './converter';
+export * from './writeback';
+export * from './utils';
